+++ pmlc/rt/level_zero/level_zero_utils.cc
@@ -2,9 +2,10 @@
 #include "pmlc/rt/level_zero/level_zero_utils.h"
 
 #include <string>
+#include <utility>
 #include <vector>
 
-namespace pmlc::rt::level_zero {
+namespace pmlc::rt::level_zero::lzu {
 #if 0
 // Returns OpenCL version as decimal number, ie:
 // OpenCL 1.2 - 120, OpenCL 2.2 - 220
@@ -34,19 +35,21 @@ static bool isPlatformSupported(const cl::Platform &p) {
   return true;
 }
 #endif
-std::vector<ze_device_handle_t> getSupportedDevices() {
+std::vector<std::pair<ze_driver_handle_t, ze_device_handle_t>>
+getSupportedDevices() {
   /*static bool first = true;
   if(first) {
     zeInit(0);
     first = false;
   }*/
-  std::vector<ze_device_handle_t> supportedDevices;
-  for (auto driver : lzt::get_all_driver_handles()) {
-    for (auto device : lzt::get_devices(driver)) {
-      supportedDevices.push_back(device);
+  std::vector<std::pair<ze_driver_handle_t, ze_device_handle_t>>
+      supportedDevices;
+  for (auto driver : lzu::get_all_driver_handles()) {
+    for (auto device : lzu::get_devices(driver)) {
+      supportedDevices.push_back(std::make_pair(driver, device));
     }
   }
   return supportedDevices;
 }
 
-} // namespace pmlc::rt::level_zero
+} // namespace pmlc::rt::level_zero::lzu
